+++ CHANGES.rst
@@ -63,7 +63,8 @@ Internal Changes
 - Replace WYSIWYG (rich-text) editor with TinyMCE, due to the license and branding
   requirements of the previous editor (:pr:`5938`)
 - Add a new Indico design system (:pr:`5914`, thanks :user:`foxbunny`)
-  fields on a per-registration basis (:pr:`5424`, :pr:`5924`)
+- Add ``event.registration_form_field_deleted`` signal, allowing plugins to handle
+  the removal of registration form fields (:pr:`5924`)
 
 
 ----